Verify the key in an apt repository
Add a facility to detect invalid or expired keys. Invoke this on the
cached key file if a uri is being used, invoke this on the current
apt-key output if a fingerprint is specified.

For remote keys, the key will be updated and reverified if the key is
found to be invalid.

For local keys, the run will fail if the key specified in the cached
file is found to be invalid.